Split controller.go off of the Labeler reconciler
