+++ pkg/reconciler/labeler/controller.go
@@ -1,0 +1,50 @@
+/*
+Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package labeler
+
+import (
+	"github.com/knative/pkg/controller"
+	servinginformers "github.com/knative/serving/pkg/client/informers/externalversions/serving/v1alpha1"
+	"github.com/knative/serving/pkg/reconciler"
+)
+
+const (
+	controllerAgentName = "labeler-controller"
+)
+
+// NewRouteToConfigurationController wraps a new instance of the labeler that labels
+// Configurations with Routes in a controller.
+func NewRouteToConfigurationController(
+	opt reconciler.Options,
+	routeInformer servinginformers.RouteInformer,
+	configInformer servinginformers.ConfigurationInformer,
+	revisionInformer servinginformers.RevisionInformer,
+) *controller.Impl {
+
+	c := &Reconciler{
+		Base:                reconciler.NewBase(opt, controllerAgentName),
+		routeLister:         routeInformer.Lister(),
+		configurationLister: configInformer.Lister(),
+		revisionLister:      revisionInformer.Lister(),
+	}
+	impl := controller.NewImpl(c, c.Logger, "Labels", reconciler.MustNewStatsReporter("Labels", c.Logger))
+
+	c.Logger.Info("Setting up event handlers")
+	routeInformer.Informer().AddEventHandler(reconciler.Handler(impl.Enqueue))
+
+	return impl
+}
